Simplify models: Move enums to core, add tracking mixin, use table args
