+++ app/models/email/template.py
@@ -8,12 +8,16 @@ from sqlalchemy import String, Text, Boolean, ForeignKey
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.models.base import Base
+from app.models.email.mixins import EmailTrackingMixin
 
 
-class EmailTemplate(Base):
+class EmailTemplate(EmailTrackingMixin, Base):
     """Email template model."""
     
     __tablename__ = "email_templates"
+    __table_args__ = (
+        {"sqlite_on_conflict": "ROLLBACK"}
+    )
 
     id: Mapped[UUID] = mapped_column(
         primary_key=True,
@@ -56,6 +60,7 @@ class EmailTemplate(Base):
 
     # Relationships
     client = relationship("Client", back_populates="templates")
+    emails = relationship("Email", back_populates="template")
 
     def __repr__(self) -> str:
         """String representation."""
